wallet/requirements.txt
amqp==5.2.0
asgiref==3.8.1
async-timeout==4.0.3
billiard==4.2.0
black==24.3.0
blinker==1.7.0
celery==5.3.6
certifi==2024.2.2
charset-normalizer==3.3.2
click==8.1.7
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
coreapi==2.3.3
coreschema==0.0.4
Django==3.2
django-redis==5.4.0
django-rest-swagger==2.2.0
djangorestframework==3.15.1
drf-yasg==1.21.7
Flask==3.0.3
flower==2.0.1
humanize==4.9.0
idna==3.6
inflection==0.5.1
itsdangerous==2.1.2
itypes==1.2.0
Jinja2==3.1.3
kombu==5.3.6
MarkupSafe==2.1.5
mypy-extensions==1.0.0
openapi-codec==1.3.2
packaging==24.0
pathspec==0.12.1
platformdirs==4.2.0
postgres==4.0
prometheus_client==0.20.0
prompt-toolkit==3.0.43
psycopg==3.1.18
psycopg2-binary==2.9.9
psycopg2-pool==1.2
python-dateutil==2.9.0.post0
pytz==2024.1
PyYAML==6.0.1
redis==5.0.3
requests==2.31.0
simplejson==3.19.2
six==1.16.0
sqlparse==0.4.4
tomli==2.0.1
tornado==6.4
typing_extensions==4.11.0
tzdata==2024.1
uritemplate==4.1.1
urllib3==2.2.1
vine==5.1.0
wcwidth==0.2.13
Werkzeug==3.0.2
